refactor objects into own file
